Test version: Navbar with mock api call + fixes
Fixed home appearing in all pages with exact prop on Route component. Also CreateBoard and BoardView were appearing at the same time so added exact prop in all Routes to be easier.

Since there are no interfaces in javascript, for now the class NavBarApi is serving an interface... maybe in the future we might use the implement.js library for this problem or use Typescript.

NavBar.js is in a test version. If modified correctly it will work but its in this state so that I can go back to this commit and show the problems that I dont understand. The topics are State, when render() and componentDidMount() are called and their relation + setState() calling render and blocking or not?? Finally why is it necessary to say line 12 - initiate state property and why navMenu property of state can't be initiated with null?!