Feed in Home page working. #11
Added home_page dir. to better group the classes and components that serve the same "feature" or help to render one page.

Still not the perfect solution but its working with mocked data. The main issues and doubts are: The HTTP request to /home should be in the component Home or in the App? Because in the Home component I have links for the navigation menu that I can't use... unless I have a way to update the parent's state.

About the NavBar, switched the api call from componentDidMount to componentWillMount because of the lifecycle methods order. Before the render() gets called, the method that gets called is componentWillMount().