Changes made to CreateBoard and BoardView working
Renamed all functions that return Promises to include the sufix "Async" for better readability, and its easier to know when we can use async/await.

Moved dummy data to a proper API function and its mock. Since the response of the API has changed, several functions were made to: map the hal+json response to a model object;