fix(server): Demo fix on createBlackboardItem regarding the user object
We need the line:
val user = usersRepo.findByIdOrNull(userId) ?: throw NotFoundUserException().
This is due to the fact that the User object we receive on the controller doesn't have all fields with the necessarý values.

Refactor on exception handlers for user. When the user is not found in the context of a login, then that is a 401 Unauthorized, since the given information is invalid. If it is in the context of UserController, then its 404 Not Found, since that resource doesn't exist.