feat(server): Added hal+forms responses + fixed some links
The outputModels have an id property so that the client has identifiers to use on the URLs.

Finished implementing pagination support on MultipleBoardsResponse. The queries property enables the client to make a request with any page number.
Issue with pagination: Page count starts at 0 and I couldn't find a way to change to 1. This generates a bug on the client!