feat(server): Sub. web app to Topics using FCM
Using coroutines to multiple subscription requests in parallel. Since we are using Spring MVC we need to block at the end (pattern Fork-Join).

Fix userProfile link since /profile endpoint didn't exist.

addUserToBoard endpoint changed since we now have a dependency to FCM. So the client needs to give its registration token to the server, in order to subscribe the app instance to topics.